Added input quantities and quantity consideration
User can now input quantity of an ingredient they have, for example, "flour - 200g", the algorithm will take into account that the user only has 200g of flour. If the user does not input an amount it is assumed they have infinite for the sake of recipe search. Issue with the amounts is that it only extracts the number right now so flour - 200g could be seen as 200 cups, or 200 lbs by the recipe search.